requirements.txt
pyspark==3.5.0
pandas>=2.0.0
Faker>=2.0.0
pyyaml>=6.0
pyarrow>=12.0.0 # Mejora performance